requirements.txt
Django==4.0.2
djangorestframework==3.13.1
psycopg2-binary
Faker==12.0.0
psycopg2-binary==2.9.3
requests==2.27.1
